+++ tt-media-server/model_services/queues/__init__.py
@@ -1,0 +1,9 @@
+# SPDX-License-Identifier: Apache-2.0
+#
+# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+
+from model_services.queues.memory_queue import SharedMemoryChunkQueue
+from model_services.queues.tt_faster_fifo_queue import TTFasterFifoQueue
+from model_services.queues.tt_queue import TTQueue
+
+__all__ = ["SharedMemoryChunkQueue", "TTFasterFifoQueue", "TTQueue"]
